refactor: 요금 패키지 분리
